+++ CHANGELOG.md
@@ -97,6 +97,8 @@ with the `YY.MINOR.MICRO` scheme.
 - [picodata#2842] Fixed the `schema version has changed: need to re-compile SQL statement`
   error, which could occur when you execute multiple DQL queries due to
   yield during cache eviction.
+- [picodata#2888] Fixed a bug where cluster would fail to bootstrap if there were no voter
+  instances in the initial `--peer` set.
 - [picodata#2732] Fixed a panic when attempting to inherit privileges via SQL
   (e.g., `GRANT admin TO somebody`). We do not support privilege inheritance
   via `GRANT user1 TO user2`. The system now validates the grantee type and